Update to fix server segment fault issue
The segment fault in server pipeline which is caused by the
last two previous commits, reported in call of Stitch() in
StreamReceiver.

This update is seemingly a build from the same code, but still
fix the issue anyway. It is recorded so that we can refer to
these information in the git log.