Shared memory allocation mode adjustment
A shared memory KeyCount list was saved to allocate shared memory,
but there were preempting happend when using multiple threads or
processes.

To avoid this issue, MainEncoder creates a circular queue of
FrameSM at its initialization step, each FramesSM holds necessary
information for the shared memory of one single Frame.

Keep adding new FrameSM to the queue when shared memory in the
head of queue is not able to free, this dynamically growing queue
controls all of shared memory.